requirements.txt
flask
gdown
pillow
piexif
torch
torchvision
